updated pickem leaderboard tbl

+++ dbt_project/models/marts/pickem_leaderboard.sql
@@ -3,16 +3,16 @@ select distinct
     p.tournament_name,
     p.user_id,
     l1.player_name as tier1_pick,
-    l1.total as tier1_total,
+    l1.total::TINYINT as tier1_total,
     case when upper(l1.pos) in ('WD', 'CUT') then 1 else 0 end as tier1_cut,
     l2.player_name as tier2_pick,
-    l2.total as tier2_total,
+    l2.total::TINYINT as tier2_total,
     case when upper(l2.pos) in ('WD', 'CUT') then 1 else 0 end as tier2_cut,
     l3.player_name as tier3_pick,
-    l3.total as tier3_total,
+    l3.total::TINYINT as tier3_total,
     case when upper(l3.pos) in ('WD', 'CUT') then 1 else 0 end as tier3_cut,
     l4.player_name as tier4_pick,
-    l4.total as tier4_total,
+    l4.total::TINYINT as tier4_total,
     case when upper(l4.pos) in ('WD', 'CUT') then 1 else 0 end as tier4_cut
 
 from {{ ref('fct_picks') }} p
